+++ src/main/java/com/telepathicgrunt/blame/mixin/BlameMixinPlugin.java
@@ -1,0 +1,54 @@
+package com.telepathicgrunt.blame.mixin;
+
+import net.minecraftforge.versions.forge.ForgeVersion;
+import org.apache.maven.artifact.versioning.ArtifactVersion;
+import org.apache.maven.artifact.versioning.DefaultArtifactVersion;
+import org.objectweb.asm.tree.ClassNode;
+import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
+import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
+
+import java.util.List;
+import java.util.Set;
+
+public class BlameMixinPlugin implements IMixinConfigPlugin {
+
+    @Override
+    public void onLoad(String mixinPackage) {
+    }
+
+    @Override
+    public String getRefMapperConfig() {
+        return null;
+    }
+
+    @Override
+    public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
+
+        if(mixinClassName.equals("com.telepathicgrunt.blame.mixin.FlatGenerationSettingsMixin")){
+            ArtifactVersion forgeVersion = new DefaultArtifactVersion(ForgeVersion.getVersion());
+            ArtifactVersion cutoffForgeVersion = new DefaultArtifactVersion("36.1.18");
+            if(forgeVersion.compareTo(cutoffForgeVersion) >= 0){
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    @Override
+    public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {
+    }
+
+    @Override
+    public List<String> getMixins() {
+        return null;
+    }
+
+    @Override
+    public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+    }
+
+    @Override
+    public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {
+    }
+}